v0.5 - paste mode and onresult callback + codedent out of the box
This version adds a `board.paste(code, {hidden:true})` helper
to simply paste huge chunks of code.

The code in both `paste` and `eval` is also now automatically indented,
so that code from the page or textarea or anything else *should* not break.

The initialization also accepts `onresult` helper which is, by default,
`JSON.parse`, but it could be `json.loads` instead to have Python references,
as opposite of JS proxies, then the board evaluated code that resulted into
dictionaries, lists or tuples.